Add files via upload
This CNN project tackles image recognition! We'll train a computer to identify objects in pictures. We feed it a labeled dataset (think cat photos labeled "cat"). The CNN, inspired by the brain's structure, analyzes patterns in the images. It extracts features like edges and learns to differentiate cats from dogs. Finally, we test the trained model on new images, and it predicts what's in them